product variant item  updated

+++ app/Http/Controllers/Backend/ProductVariantController.php
@@ -6,6 +6,7 @@ use App\DataTables\ProductVariantDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\ProductVariant;
+use App\Models\ProductVariantItem;
 use Illuminate\Http\Request;
 
 class ProductVariantController extends Controller
@@ -94,6 +95,10 @@ class ProductVariantController extends Controller
     public function destroy(string $id)
     {
         $productVariant = $productVariant = ProductVariant::findOrFail($id);
+        $variantItemCount = ProductVariantItem::where('product_variant_id', $productVariant->id)->count();
+        if($variantItemCount > 0 ) {
+            return response(['status' => 'error' , 'message' => 'This variant has items in it delete the variant item, firt for delete variant item']);
+        }
         $productVariant->delete();
 
         return response(['status' => 'success', 'message' => 'Deleted Successfully']);
